refactor: add standalone test module
fixes #2957
