+++ src/CMakeLists.txt
@@ -52,6 +52,7 @@ endfunction(compile_lib)
 
 function(compile_test DIR)
     set(TEST_LIBS
+        openmldb_test_base
         apiserver nameserver tablet query_response_time openmldb_sdk openmldb_catalog
         schema client zk_client storage replica base openmldb_codec openmldb_proto log
         common zookeeper_mt tcmalloc_minimal ${RocksDB_LIB}
@@ -96,6 +97,9 @@ compile_proto(sql_procedure ${PROJECT_SOURCE_DIR})
 compile_proto(api_server ${PROJECT_SOURCE_DIR})
 compile_proto(taskmanager ${PROJECT_SOURCE_DIR})
 
+add_library(openmldb_proto STATIC ${PROTO_FILES})
+add_library(openmldb_flags STATIC flags.cc)
+
 compile_lib(openmldb_codec codec "")
 compile_lib(openmldb_catalog catalog "")
 compile_lib(schema schema "")
@@ -113,10 +117,6 @@ compile_lib(apiserver apiserver "")
 find_package(yaml-cpp REQUIRED)
 set(yaml_libs yaml-cpp)
 
-add_library(openmldb_proto STATIC ${PROTO_FILES})
-
-add_library(openmldb_flags STATIC flags.cc)
-
 set(BUILTIN_LIBS apiserver nameserver tablet query_response_time openmldb_sdk openmldb_catalog client zk_client replica base storage openmldb_codec schema openmldb_proto log ${RocksDB_LIB})
 set(BIN_LIBS ${BUILTIN_LIBS}
 common zookeeper_mt tcmalloc_minimal
@@ -126,6 +126,7 @@ ${ZETASQL_LIBS}
 ${BRPC_LIBS})
 
 if(TESTING_ENABLE)
+    add_subdirectory(test)
     compile_test(cmd)
     compile_test(base)
     compile_test(codec)
